requirements.txt
Flask
Flask-Login
Werkzeug
numpy
pillow
scikit-image
scikit-learn
opencv-python
joblib
tqdm
